Refactor app and routes
